Make/blueMSXlite/linux-sdl/Makefile-em Src/Emulator/Emulator.c Src/SoundChips/OpenMsxYM2413_2.cpp Src/SoundChips/OpenMsxYMF278.cpp Src/Utils/IniFileParser.c

+++ blueMSX/Src/Utils/IniFileParser.c
@@ -207,7 +207,7 @@ IniFile *iniFileOpen(const char *filename)
     if (iniFile != NULL)
     {
         iniFile->isZipped = 0;
-        iniFile->isZipped = NULL;
+        iniFile->isZipped = (int)NULL;
         
         iniFile->modified = 0;
